Move reused components into /components
- Prefer .jsx extension for React
