+++ views/Admin.react.js
@@ -33,7 +33,7 @@ import { emphasize } from '@material-ui/core/styles/colorManipulator';
 
 import { compose } from 'redux';
 import { withStyles } from '@material-ui/core/styles';
-import DrawerComponent from './Drawer.react';
+import DrawerComponent from './components/Drawer';
 import Drawer from '@material-ui/core/Drawer';
 import Hidden from '@material-ui/core/Hidden';
