+++ views/Subject.react.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
-import GraphFactory from './GraphFactory.react'
+import GraphFactory from './components/GraphFactory'
 import { connect } from 'react-redux'
-import Drawer from './Drawer.react'
+import Drawer from './components/Drawer'
 import Modal from './Modal.react'
 import io from 'socket.io-client'
 import 'whatwg-fetch'
