+++ views/components/AbstractGraph.d3.js
@@ -1,5 +1,3 @@
-import * as d3 from 'd3';
-
 export default class AbstractGraph {
   constructor(el, props) {
     this.el = el;
