+++ views/components/Drawer.jsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import { render } from 'react-dom';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
