+++ views/components/GraphFactory.jsx
